Move config from `learningpathApi` to `myndlaApi`
