+++ src/api/learningpathApi.ts
@@ -10,7 +10,6 @@ import {
   ILearningPathV2,
   ISearchResultV2,
 } from '@ndla/types-backend/learningpath-api';
-import { IConfigMetaRestricted } from '@ndla/types-backend/myndla-api';
 import { GQLLearningpath, GQLMeta } from '../types/schema';
 import { fetch, resolveJson } from '../utils/apiHelpers';
 
@@ -77,17 +76,3 @@ export async function fetchLearningpath(
     learningsteps,
   };
 }
-
-export const fetchConfig = async (
-  configKey: string,
-  context: Context,
-): Promise<IConfigMetaRestricted> => {
-  const response = await fetch(`/myndla-api/v1/config/${configKey}`, context);
-  const config: IConfigMetaRestricted = await resolveJson(response);
-  return config;
-};
-
-export const fetchExamLockStatus = async (
-  context: Context,
-): Promise<IConfigMetaRestricted> =>
-  fetchConfig('MY_NDLA_WRITE_RESTRICTED', context);
